gschem: Allow attribute value column in component selector to shrink
Symbols with long values are common, and we don't want having viewed
those to force a h-scroll-bar on symbols with short valued attributes.

If we don't allow them to shrink, the columns retain the width of the
largest item they displayed, and we often get stuck with a scroll bar
 after displaying components with lengthy attribute values.

When updating the model, invalidate the column width for the attribute
value column so it is re-sized based on the new data being shown.

We might also consider invalidating the attribute name columns,
however that gives an inconsistent column division when swithing
between symbols, which doesn't look nice. For now, assume that
the name column can keep the max width gained whilst previewing.