gschem: fix select state when double clicking an object
After the o_edit() call in x_event_button_pressed() the state is
STARTSELECT.  If a MODAL dialog is called (arc or change picture) the
state will will still be STARTSELECT after closing the dialog.

This state turns into the MOVE state right after the dialog is closed
and the first motion is done in gschem (x_event_motion()).

This is not correct. The state needs to be reset to SELECT right after
the o_edit() call.