General portability cleanup of scripts.
Expand the autoconf tests for awk to make sure that we have picked up and
awk that is good enough for the scripts that use them.  Now use this result
in various scripts instead of hard coding 'gawk' (not available on all systems)
or 'awk' (picks up the original awk which is not sufficiently advanced on solaris).

Also avoid calling out /bin/bash on scripts for which /bin/sh (even on solaris) should
be sufficient.  Not all systems have /bin/bash.

Finally, improve temp file usage in shell scripts for improved security.