Remove some old comments better explained in commit history
Comments annotated with a dates and developers initials clutter the
code with historical details people may be reluctant to remove. Not
all of these are still relevant, and those which are should probably
could be kept in such a way as not to suggest code ownership. The
commit revision log is a better place for determining authorship.

Fix some whitespace in surrounding areas where I couldn't resist.