libgeda: Check for NULL before updating slots after attaching attribs
If reading a file which contains an illegal object type (which fails to
read), then contains syntax attaching attributes to that object, we may
attempt to update the slot on a NULL, or incorrect object.

Updating slots on the wrong object is harmless, but causes a crash for
NULL objects. Check for the NULL pointer before doing the update.

Also check for NULL before handling embedded object data, and set
new_obj to NULL when we read an invalid object.