gattrib: Fix crash when removing attributes from multiple files.
We were iterating over a list which could have objects removed from it.
Since this is dangerous, iterate over a copy of the list. Since
attributes come later in the object list than the objects they are
attached to, iterating the list backwards means we don't
encounter objects we've already free'd.

We also need to set TOPLEVEL->page_current to the page we're working on,
since within the loop, this variable is used to find which linked list
to remove the objects from.