Use (rubber_visible == 0) to avoid "undrawing" invisible objects.
For some operations, such as text and component placement, there is a
gap between inserting components into the place buffer (at 0,0) and
knowing where to draw them.

This is presently handled with separate one-shot states used to special
case the first mouse event recieved. These states are transitioned from
as soon as the first drawing operation has been made. If we track the
fact that no "undrawing" should take place, it's otherwise OK to use the
same motion event handler for the first and subsequent motions.