libgeda: Clean up o_*_shortest_distance() functions
Make all functions take an OBJECT * as their first argument, rather
than a BOX, CIRCLE etc.. This keeps the function signature uniform
and allows the routines full access to the whole OBJECT state.

Also use g_return_val_if_fail() as elsewhere in the code. Favour
primitive types such as double and int, rather than gdouble and gint
to be consistent with the majority of the gEDA codebase.