gschem: Make key press / release events update modifier states
If a sensitive action is taking place when a modifier key changes state,
fetch the current mouse pointer and call the action's motion handler.

This allows us to see instant on-screen changes when (say) pressing the
modifier key to switch magnetic net mode off, or between orthogonal /
non-orthogonal constraints mode when copying or moving objects.

Preivously motion was needed after pressing the modifier key to get an
on-screen update.