libgeda: Fix complex_parent for prim_objs of embedded objects
When we read in the embedded objects for a complex, we need to
set the complex_parent argument for each of them to point at the
complex.

This bug manifested in failed netlisting for embedded objects,
and has been manifest since:
commit f4dab83b1ca0e197f02f520a6ae525aeac5a252d
gnetlist: Use the complex_parent property of an OBJECT, not the HEAD node.

The omission which made this a latent bug was in:
commit ec2e929742bb682be52d02818916b1f57c343e48
libgeda: Set complex_parent on all prim objects, not just their HEAD node