o_move_end_rubberband(): Rework creating connectivity lists. Fix #2017356
We were previously making a lists of pre- and post-move connected objects
incrementally as we moved each individual object being stretched.

In some cases this meant that as we moved one object, we artifically
changed the connectivity of future ones to be moved, making the
resulting connectivity lists inaccurate. In some examples, the objects
being stretched themselves appeared in the lists.

We want to avoid this, since if we delete a stretched object when
it reaches zero length, we don't want that object to be referenced in
the list of connected objects. Since we don't want to do more work
than necessary re-drawing affected objects on screen, it is generally
desirable to keep the lists as accurate as possible.

To avoid the problem with incremental updates, build the connectivity
lists before and after the entire move operation for stretched objects.

This alone doesn't fix the case where a stretched object being deleted
happens to touch another to start with. We have to check when deleting
an object for its presence in the prev_conn_objects list. We check and
remove _all_ references to the deleted object from that list.

Thanks go to Patrick Bernaud for the test-case for this very tricky to
reproduce bug, and an initial patch.