gschem: Avoid critical assersion being triggered when cancelling move
We need to set inside_action to 0 before calling o_undo_callback(),
since that then calls o_redraw_all(). We don't want that redraw to
think we're still inside an action and attept to draw the (now empty)
place list. For good measure, place that call after we've cleared out
the stretch list as well.