gschem: Add size of grip (or cue) to invalidate and redraw regions.
Adding to the invalidated region cures screen artifacts left due
to the bounds of an object not including the area occupied by its
rubberbanding grips.

Adding to the coordinates used with s_page_objects_in_region() ensures
that grips / cues from adjacent objects are drawn correctly when repainting
a damaged region from the expose handler.

This is a bit of a hack and could be better catered for by invalidating
the grip regions separately from an object specific invalidate function.
s_page_objects_in_region() would also have to return objects with grips
or cues touching the passed region.