libgeda: functions to get the position of an object
added a function o_get_position() to get the world position
of an object.

The returned position is dependant on the object type.
For the box and picture object the lower left corner is used.
For line, net, bus, path, the first point is used.
For pin the whichend is used.
For circle and arc the center is returned.
Complex and text object return it's base coordinate.