gschem: Remove stretched object from stretch list if we delete it
When a stretched object becomes zero length when rubberbanding, we
delete it. We must also remember to delete it from the stretch list,
otherwise it will be referenced later, and could cause a crash.

NB: This isn't the whole story.. the before / after connectivity
    lists can still reference the deleted object and cause a crash.

(based on commit fe8640898cb843b72e1c3cc01ee52c33db736ccf)

Edited since the STRETCH list is not a GList in the 1.4.x series, and
other cleanup work on this file is not present.