Add explicit 'enabled' field to COLOR structure.
Colors are implicitly 'enabled' by having values specified in the
configuration file. Up until now, whether a given color is enabled has
been checked by checking if the color name strings are NULL.

This patch adds an explicit 'enabled' field to the COLOR structure. It
is set TRUE when s_color_request() is called for that color index.