Fix attribute colours for embedded objects [#2157598]
Add an argument, "set_color" to o_attrib_attach[_list]() to specify
whether the colour of the attribute should be reset to the default
attribute colour when performing the attachment.

We set this to true when promoting toplevel unattached attributes
inside a symbol onto the schematic page, when the user explicitly
attaches an attribute inside gschem. It is FALSE in other cases.

In cases where we're creating new attributes, such as o_slot_end(),
we specify the attribute colour we want to o_text_new().

Remove colour setting from o_attrib_add(), since there is no point
having this code in two places. o_attrib_add() just deals with the
datastructures now.