gschem: Separate out dotted grid drawing and tidy up the code somewhat.
Creates a separate static function for drawing the dotted style of grid,
and a switch function to pick which grid drawing routine is in use, which
is currently just dotted grids or nothing.

Adds doxygen documentation for the grid drawing functions.

Renames the "grid-color" colour map index and other dotted grid specific
config parameters to prepend "dots-".