gschem: New helper function to query rendered grid spacing.
The status bar contains an label which lists the current rendered grid
spacing. We should not be forcing an update of this each time we repaint
a portion of the grid. Introduce a helper function to ask the grid code
what size it would end up drawing. This will then allow us to query the
rendered grid size after zoom / grid settings changes, and update there.