gschem: Ensure we don't allow the preview to load from file _and_ a buffer
When setting either the "filename", or "buffer" property on the preview
window, we g_free() the old values for filename and buffer. We missed the
required assignment of the old value back to NULL.

The code isn't strictly correct though, as when changing a property
internally like this, we should really emit a "notify" signal on the
property we've changed using g_object_notify().