gschem: Attempt better hinted alignment when drawing arcs.
gschem's canvas can have a fractionally different X and Y scale factor
in some cases, which can cause misalignment of arcs drawn assuming the
scale factors are uniform. This manifests because at high zoom levels,
the center of an arc can be a large distance from the drawn stroke,
over which distance the scaling errors magnify to give a visible offset.

The other adjustment made here is rather than hinting the center of the
arc to the pixel grid, we take the case where the arc forms a full circle,
then find the center (in floating point screen coordinates) which would
align the tangent of the arc at 0, 90, 180 and 270 degree positions, to
lines hinted to the pixel grid at those positions.

A 0.5px offset correction made in gschem_cairo_arc() to the arc center,
for odd stroke width accounts for lining up the those tangents with
the hinting which is applied to odd width lines in gschem_cairo_line().