gschem: Adjust hinting of lines to include an extra overlapping pixel
When drawing using GDK, a span of pixels being rendered in a line would
light all pixels touched by the span.

Since moving to cairo, our hinting code has rounded the line's end-points
to land between pixels. (So cairo paints the whole pixel), however it
means that there is one pixel less being painted (as both ends are
adjusted in the same direction).

Fix this with a manual addition of 1px as appropriate to the coordinates
such that an extra pixel is hit for horizontal and vertical lines.

This now means that a line drawn between two grid positions will acually
be drawn on top of both of those grid lines, rather than being drawn on
top of one, and just touching the other.

Apply the offset factor (0 or 0.5px) to both X and Y coordinate pairs for
non horizontal / vertical lines, since that gives better alignment with
the grid.

Apply the same hinting to round capped lines as to square capped.