gschem: Group fill hatching into a single cairo path before stroking
This ought to improve performance. Since the hatch lines are drawn
in parallel, the path doesn't self-intersect, and we can gain a speed
increase by sending the whole operation in one go. (Self-intersection,
e.g. in the case of a grid would slow cairo down computing the mask
to render with).