gschem: Ensure circles are well hinted to the grid
gschem's canvas can have a fractionally different X and Y scale factor
in some cases, which can cause misalignment of circles with other
objects. Since logic bubbles are often placed to touch the boxes of
components, ensuring they align well is important.

The code used here is similar to that added for arcs in
commit 508ab0fd9a1e4317ee660bc9eba79fcb49329841