gsymcheck: checks for overbars and escaping
Test for single overbar markers '\_label' and print a warning.
libgeda currently closes an open overbar at the end of the string
but a different implementation could drop the overbar.

It also tests for single backslash characters. They will
be invisible as a '\x' combination will only print the second
char 'x'. The user should decide whether he likes to remove
the backslash or whether to escape it and write '\\x'