gschem: Add interface to toggle a pin between net pin and bus pin types.
This allows schematic diagrams of hierarchical connections to include
buses. Since gnetlist does not currently support buses, this feature is
useful only for diagrams, or when used with a custom netlister.

Since we don't want to mislead users into thinking bus pins netlist,
the option to set pin type is only present on the page's popup menu,
and is marked "(graphical)".