gschem: Adjust position hinting when drawing round junction cues
Apply hinting to the cue's position to match the position of the object
it is sitting on. This is done by passing the on-screen width (pixels)
of a net / bus into gschem_cairo_arc(). If the line-width is odd, that
function will shift the origin of the arc by 0.5px. Refactor the places
where we draw circular arcs to use a single helper function in o_cue.c