gschem: Refactor partially duplicated drawing code in various files.
For each object type, there is drawing code for the normal object
on the page, an object being placed, and objects being rubberbanded.

In many cases (arc, circle and path being good examples), each of these
functions repeat non-trivial amounts of very similar setup and drawing
code. Only the line-width, stroke style need to be specific to each.

Refactor the code such that the place and rubberbanding routines (which
adjust the rendered geometry based on state variables) construct a
temporary object with the modified geometry. Common parts of the cairo
setup and path generation can be combined and shared between routines.