gschem_cairo_box(): Allow passing line_width as -1 to switch off hinting
This may be useful for drawing a box path which is to be filled. If at
some point, gschem_cairo_* were to take 0 line-width as instruction to
draw "hardware" type lines, passing 0 to avoid hinting would clash with
that feature.

When the line_width is passed as -1, an extra pixel is added to the
larger X and Y coordinates, so the fill touches the pixel which would
be lit if a line was drawn at the passed coordinates.