gschem: Make gschem_cairo_* functions take parameters in world coordinates.
This changes the prototype of all gschem_cairo_* functions to take a
GSCHEM_TOPLEVEL parameter rather than cairo_t as their first argument.

This change cuts down a significant number of lines of code, and also
paves the way to having world-coordinates passed straight into the cairo
routines, should we want to use cairo for printing at some stage.

A few cases of direct cairo usage in screen coordinates remain; these are
in o_cue.c, and x_grid.c. The case in o_cue.c remains because it relates
to a filled box. gschem_cairo_box () hints the path coordinates for a
stroked box and clamps the minimum on-screen stroke width to 1px, thus
rendering it unsuitable for hinting a path to be filled.