gschem: Clamp minimum size to draw stoke dots using cairo, apply hinting (Rest!)
Activate the code partially committed in the last commit, and add the
remaining code to hint the position of dots in a stroke.

Since cairo uses antialiased drawings, a dot of diameter 1px will actually
render ligher than it would have with GDK. Once the line width reaches 1px
switch back to drawing dots as squares.

Also add hinting to the position of the drawn dots drawn along the stroke
to ensure they land centered on pixels.