gschem: show correct properties in filltype dialog
The filltype dialog uses the selection list now. With the new
functions selection_get_fill_type() and fill_type_dialog_set_values()
it will be possible to convert the dialog into a non_modal dialog.

In the dialog only those values are marked as *unchanged* that
differ from each other.

This also fixes the bug [#2344964].