gschem: fill type dialog: improved default values
If the user changes the fill type from HOLLOW to HATCHED and does not
define the angle and pitch, then the gschem could either reject
the change of the type, or guess the pitch values.

I've decided that it is better to set the missing values.

In the opposite case, when the user switches the type from HATCHED to
HOLLOW, the pitch is still set. Reset the not required parameters
to -1.

Maybe the functionality should be moved to o_set_fill_type() in
libgeda.