gschem: Calculate radius of centered arcs and boxes with 2x precision
Since we may have to hint the radius by 0.5px to bring it onto the
pixel grid, we will get better results if we compute the radius to
half-pixel precision.

Re-enable radius hinting for centered arcs back on, but only for arcs
which have radius > 1px to start with. If we don't, there is an abrupt
transition between when the arc lights a single pixel, and when it is
the next largest size. A little anti-aliasing in that case helps.