gnetlist: Avoid depending on GHashTable ordering for output consistency
If the GLib hash table changes (which it has in late versions of GLib),
the netlist output ordering changes due to our mis-use of GHashTable to
list all unique component names without enforcing any eplicit order on
the enumeration of the hash table contents.

Since ordering may be important to backends, want to have known output
ordering whilst removing duplicate entrys, but without an explicit sort
of the refdes list.

This commit introduces an incremental build of the list, first checking
the hash table for a hit, and only adding the new refdes if it wasn't
already in the has table. This is gives the output a predictable order
(based on the input), rather than inserting all refdes into the hash
table, and building the list in whatever order the table implentation
traverses them in.

This change does unfortunately means a large number or ordering related
changes are also made in the gnetlist test suite's golden files.

There are also some non-ordering changes in the checked in golden files.
These is encountered in the pretty well obsolete PCBboard backend - which
errors out when it hits certain components it can't find footprints for.
Changing the ordering changes the point at which it dies, and thus the
netlist output. Lets just accept these changes as is.