gnetlist: Teach netlist traversal code to understand buses (still disabled)
Up until recently, the connection scanning code has rejected some, but not
all buses. This means that it is unlikely graphical buses formed a useful
part of anyone's design, since the netlist code in s_traverse_net()
would not follow connections to another bus segment. The only possible
bus-pin connectivity has ever been "pin-bus-pin", or "pin-bus-net-...".

This behaviour appears to have been accidental, and would not have been
easy for a user to take advantage of. The traversal code would not move
on to a subsequent bus, but would, for example, allow a pin->bus->!bus
connection, which could then connect onwards. The bus mixed up in this
would for all intents and purposes behave as a net.

The pin->bus validity bug was actually fixed in libgeda when adding
support for bus pins. This commit formalises it further, since if
support for netlisting bus pins were enabled it would otherwise still
be possible to get a buspin-bus-net-... connection if you tried hard.

Introduce type-checking between pins, nets and buses such that they can
not netlist where the object types don't match. For buses, the code to
name a connection based on "netname", "net" or "label" attributes is
disabled - pending decision as to what constructs we wish to use for
naming buses. Add a gnetlistrc setting (unnamed-busname ...) which sets
the default naming (if the connection scanning code were run on bus pins).

This commit lays some of the ground work for shared connection scanning
code for determining the connectivity of buses and bus pins.