libgeda: Avoid overlinking against libraies we don't actually use
The only libraries we need to specify for libgeda is -lgeda, and an
appropriate search path.

For CFLAGS, we need to specify the path to libgeda's include files.

We must also reference the libraries we depend upon being installed
for our API / header files. This is currently guile and gdk-pixbuf.
Since difrerent guile versions have different pkg-config files, this
is not particulary simple to achieve. To avoid the problem, lets
just not pass any guile CFLAGS on from libgeda's pkg-config file.

Libgeda requires its callers to have correctly initialised guile, so
imposing the requirement that guile's include files are available is
not a great extra burden. (They must be available for the app to be
calling the correct initalisation functions).