libgeda: Correction to the last commit message; do pass on guile CFLAGS
Depending on where guile is installed, passing along the CFLAGS might
be useful. In the future, when we can depend on guile-1.8 or later, we
could just add guile-1.8 to the Requires.private: line of libgeda.pc.in.

No code changes.. the guile CFLAGS were not dropped as described in
commit 2742d7a50cbf126ea5a901e2f9faa6259263ae75.