libgeda: Fix printing of colours.
To avoid emitting a colour change for each object printed, libgeda caches
the last colour it wrote into a postscript file, and doesn't emit a new
colour change command when the next object is the same colour.

Drawing of complex and text objects are wrapped by a "gsave" and "grestore"
pair, which resets the postscript interpreter's current colour, so we need
to restore our variable tracking the last drawn colour after these groups.

Fixes mis-drawn colours noted by DJ on the geda-user list.