Write logs to $HOME/.gEDA/logs/.
Rewrite s_log_init() to put log files in a centralised location. A
user application (e.g. gnetlist) provides a prefix, and log file is
opened at:

   $HOME/.gEDA/logs/<prefix>-<YYYYmmdd>-<serial>.log       [1]

The serial is found by the following procedure:

1. Read through the directory to find the highest existing <serial>.

2. Starting with that number, repeatedly attempt to open()
   incrementing files with O_CREAT|O_EXCL flags to guarantee that a
   new file is created.

3. If after LOG_OPEN_ATTEMPTS = 5 attempts a unique filename hasn't
   been found, fail.

The logic *intentionally* does not fill in gaps in the sequence -- if
I delete logs, I still want a new log to appear at the end of the
list.

[1] The actual location is a subdirectory "logs" of the user config
    directory, which may not be ~/.gEDA on all platforms.