tragesym: put attributes on top if there are no top pins
If you put pins on either the top or bottom of a symbol, tragesym
moved the name around and changes the positioning. However, this is
only necessary if pins are on top, and the logic should ignore pins on
the bottom.

Fix suggested on gEDA mailing list by Matt Ettus <boyscout@gmail.com>