libgeda: Remove o_attrib_search() function
This function only had one caller, which has been modified.

When attaching an attribute, we check if it is already attached.
Previously this code used o_attrib_search() to check by object
SID. Instead, we now just look for the OBJECT* pointer in the
attribs list using g_list_find.