libgeda: Add special case slot update when copying a slot= attribute
This update is needed to ensure the pinnumbers on the complex are updated
correctly if a slot= attribute is copied along with the symbol.

This hasn't caused a problem since when copying the complex, the pinnumbers
are reset to match that of the object being copied - looking at the slot=
attribute of the complex being copied.

If its slot= attribute is not copied along with the complex, the new
complex will have the incorrect pin numbers. We need to make sure that
when the slot= attribute is re-attached, it updates the pinnumbers
properly - we can then remove the pinnumber update in o_complex_copy().