Call o_attrib_slot_update(), not o_attrib_slot_copy() in o_complex_copy()
When the appropriate slot= attributes are re-attached to the complex,
the slot will be updated. There is no point in pre-empting the pinnumber
update by looking at the slot= attribute of the source complex.

If the complex were copied, but its slot= attribute were not, the call
could result in incorrect pinnumbers on the new object.

We do want to run o_attrib_slot_update() though, just in case the symbol's
internal slot= attribute would cause a pinnumber update.