Point at the correct gnetlist and scheme files for all tests.
When running the tests, point at the not-yet-installed scheme files.
This lets 'make check' test out what you just built instead of something
that is installed that may be sorely out of date.

Point at the locally built but perhaps not yet installed gnetlist
when running 'make check'.  This helps make sure we are testing
the right gnetlist.

While here also improve the shell portability a little by using

  FOO=value
  export FOO

instead of

  export FOO=value

which is not quite as portable.