Add inherited attributes to the list returned by o_attrib_return_attribs()
This starts to allow a common place where attributes from the symbol will
be combined with attached attributes to form an aggregate attribute list.

Currently, various pieces of code which deal with reading attributes will
check the attached attributes of a symbol, then the inherited attributes
explicitly. For cases where a simple combining rule exists, we can reduce
such checks to a single lookup of the aggregate list.

Atttributes such as "net=" and "source=" which often exist multiple times,
possibly both as inherited and attahched attributes, present difficulties
though.

For now, just pile all inherited attributes at the end of the returned
list. This means any code which wants to check for the first attribtue
with a given name should do the "right thing", with attached attributes
taking presidence over inherited ones.