Introduce new attribute searching API
This new API aims to be explicit as to the scope of its behaviour.
Various of the existing APIs are very similarly named, yet will have
subtly different behaviours when passed a list of OBJECTs, attribute
OBJECTs, some general OBJECT, or an attribute OBJECT.

Since the current APIs are confusing, the sensible course of action seems
to be introducing a new simple API for searching a GList of attributes by
name, then providing specialised wrappers which construct the appropriate
list of attributes to search.

To this end, inside libgeda/src/o_attrib.c, the following new functions
are created:

The generic search routine:
  static OBJECT *o_attrib_find_attrib_by_name ()

A helper which returns the value of the found attribute:
  static char *o_attrib_search_attrib_list_by_name ()

Exported API to search:

  floating attributes (on a page, or inside a symbol):
    char *o_attrib_search_floating_attribs_by_name ()

  Attributes attached to an OBJECT:
    char *o_attrib_search_attached_attribs_by_name ()

  Attributes inside an OBJECT (inherited):
    char *o_attrib_search_inherited_attribs_by_name ()

  Attributes from an object - either attached, or from inside the symbol:
  char *o_attrib_search_object_attribs_by_name ()


The latter is in preparation to make more use of an aggregate attribute
list for objects - rather than having each piece of code duplicate the
logic to check both attached attributes, and those inside symbols.

We also make o_attrib_find_floating_attribs() static, since it is
only used inside o_attrib.c now.