Move logic of function o_attrib_search_toplevel_all() to its one caller
This function was only used in one place, and probably didn't deserve its
own API in libgeda. The function it is called from is not used anyway, so
could probably be removed completely at some point. (Although it forms
part of the scheme API for gnetlist, comments above it suggest that it is
"still highly temp and doesn't work right"). It is not requied by any
known (shipped) gnetlist backends.