libgeda: Eliminate usage of, and remove o_attrib_search_string_partial()
This function's one caller was using it to find an attribute containing
"slotdef=X:", where X is the slot number being retrieved.

o_attribute_search_string_partial() would look at the full name=value
string of each text object in the given object list, looking for a
substring match on the desired text. (In fact, we only want a match
at the beginning of the string).

The new implementation may be somewhat less efficient, since it queries
all slotdef= attributes in turn, checking to see if the one returned is
the desired value. The speed penalty should be impercievable though, and
the new code better abstracts the implementation details of the attribute
system.