Replace o_complex_return_pin() with new o_complex_find_pin_by_attribute()
Implement a more generic version of o_complex_return_pin(). The old
function takes a pinnumber as its argument, and finds a pin which has
a matching "pinnumber=..." attribute.

The new function takes an explicit attribute name argument as well as
the value, so we can find pins by any attached attribute. This is useful
so we can search by "pinseq=..." attributes as well.