libgeda: Use new attrib API for o_attrib_slot_update()
Also tidy up the function a little.

We now use o_attrib_search_object_attribs_by_name() to find the
slot= attribute value of an OBJECT, so we automatically get the
correct fallback to the symbol's embedded default slot= attribute
value if the attached slot= attribute is missing.

(This means a behavioral change in the uncommon case where the
attached slot= attribute is missing, and the symbol specifies a
default which is not slot=1)