gschem: Deselect invisible attribs with their parent object (Bug #2430369)
Previously, when a component was selected and you deselected it with a
ctrl-click or ctrl-drag, the invisible attributes attached to the
component remained selected.

This has caused problems for users in the past, when trying to delete
everything in a given area - except for a few components in the middle.

Drag selecting the area, deselecting the desired components, then deleting
would also delete the (still selected) invisible attributes belonging to
the deselected components.

To fix this, make sure that when we remove an object from the selection
list, we also remove any invisible attributes. When invisible text is
being shown, there is no need to do this, since the user can manually
de-select these attributes. (And they may want the finer grained control
afforded by the ability to do so).