Add GschemAction and GschemAccelLabel subclasses of their Gtk equivalents
These subclass are intended to encapsulate information relating to
gschem's multi-key shortcuts. Their code is based on GTK 2.14.5 (LGPL).

For GschemAccelLabel, we have to subclass GtkAccelLabel rather than just
GtkLabel, otherwise the menu doesn't correctly pad the accelerators into
their own column.

To set the accel_width, we have to go poking in a private part of the
GtkAccelLabel widget. The relevant field has been protected with GSEAL,
so this will not work in GTK3.0, or when compiled with GSEAL active.