gschem: Add "_" style accelerators within menus
Copying accelerators for other common applications, such as "_Open",
"_New" etc. This aids keyboard navigation once we are in the menus
(which have a key-grab, so gschem's multi-key accelerators don't work.)

To set the accelerator on "Open Recen_t", we also have to edit x_menu.c,
since the recent files magic is hooked on the menu text (including the
accelerator). We also remove the explicit override of the label text in
x_menus.c, since we want to show the accelerator.