libgeda: Don't use intltool.
intltool causes a number of problems at the moment. It's not
tremendously portable, it hammers the gettext infrastructure heavily,
and it has large incompatibilities between versions.

Instead I've created a tool called desktop-i18n which does all of the
things we currently use intltool for, without any dependencies except
for a working shell & make. It's not as featureful as intltool, but it
appears to do everything that is required.

Thanks to Dan McMahill for testing and portability suggestions.