Fix the testsuite when building outside the source tree.
Since gnetlist currently only looks in a single directory for scheme
files and since some of our scheme files are generated at build time
we need to copy over the other source .scm files to the build directory
for the testsuite to run.  Longer term, we should either

a) consolidate all of the testsuite into a single master script that
   starts by merging all the scheme files into a single directory

or

b) teach gnetlist to have a scheme search path.  Even with this,
   testsuite consolidation would be good.