Quote pins names if they contain comma or close parenthesis. [2823703]
A comma or close parenthesis will cause problems with the pcb action
script, so if one of the arguments to ChangePinName contains one it
should be quoted.  Any quote characters within the argument are
escaped.

Right now it only quotes if there is a comma or close parenthesis. We
could just quote all pin names to future proof it against more special
characters in the action argument, but this seems less intrusive and
doesn't affect those without a new pcb that accepts quotes in the
arguments.

(Minor formatting changes and addition of comments made by Peter B
prior to committing).