libgeda: Fix font loading.
Fixes a bug introduced in 8af5fba2ec0248331ceb5d2746771e70b2813004
which caused libgeda to not detect character widths correctly.

This was subtle.  A new font OBJECT must be created and added to the
global character hashtable *before* o_read() is called to populate the
OBJECT with its constituent lines.  This is because when o_read()
encounters the 'F' font parameter line, it calls
o_text_set_info_font() to process it.  o_text_set_info_font() in turn
needs to be able to get at the font object for that character, which
it looks for using the global hashtable.

I have added some comments to the source code explaining this.