Trimmed the filename length of some gxyrs output files
make distcheck was having some path length issues with the renamed files
in this commit.  More specifically, tar was complaining the path length
exceeded some sort of limit, which eventually caused make distcheck to fail
badly.  This bug was ticked by the next commit (where the package name was
lengthened).