Renamed the package name/tarball to be more accurate
I don't think we should call the make dist created tarball "gEDA-VERSION..."
since technically gEDA is more than just gEDA/gaf (it also consists of
PCB, Icarus, gerbv, gnucap, ngspice, etc...).   I think that
gEDA-gaf-VERSION... is a fine name.  I briefly toyed with geda-gaf-VERSION...
but geda is just the wrong way of writing gEDA (however, if there is a
platform/filesystem reason why it should be all lowercase...).

Also, as much as I would like all the gEDA projects to share a single
configure script and/or git repository, that is just not going to happen.
Unification and/or creation of the entire suite is a different world/task
than the toplevel gEDA/gaf configure.ac file.