If pkg-config fails to find guile then fall back to guile-config.
guile-1.8.0 is the current minimum required guile version for gaf.
guile-1.8.5 is the first version that shipped with a pkg-config file
so fall back to guile-config if we fail to locate guile with pkg-config.
In addition, if we fail to find guile-1.8 with pkg-config check for
guile-2.0 with pkg-config.  Per a discussion with guile developers they
anticipate shipping a guile-2.0.pc.in file when gule-2.0 comes out.