build-sys: Improve HTML conversion rules.
We want to make sure the HTML version of the man pages is always
distributed, but make building/installing it optional (in case users
don't have a proper groff installed).

This can be achieved by making sure that Automake only tries to
install the HTML man pages if a working groff is available.  In
addition, `make distcheck' is set up to behave as if groff is not
installed.

Unfortunately, this still isn't ideal, because users who build from a
source archive (with pre-built HTML) won't get HTML installed unless
they have a working groff.

In the long run, what needs to happen is to use a mechanism like
`missing' to handle rebuilding of HTML man pages -- or just drop HTML
man pages altogether.